Moved get_datetime_from_timestamp out of riot
