+++ src/riot.py
@@ -1,4 +1,3 @@
-from datetime import datetime
 import requests
 
 from . import configuration
@@ -99,13 +98,3 @@ def get_champion_names_from_ids(ids, champions, release=settings.latest_release)
     champ_ids_to_names = get_champions_map(champions, key="key", value="id")
     names = [champ_ids_to_names[id] for id in ids]
     return names
-
-
-def get_datetime_from_timestamp(timestamp):
-    """
-    From the doc: https://riot-api-libraries.readthedocs.io/en/latest/specifics.html
-    "The creation date timestamps in milliseconds (not seconds)."
-
-    Returns: datetime object
-    """
-    return datetime.fromtimestamp(timestamp / 1000)
